various bug fixes, hiding chat to only 1 to 1 chat

+++ src/java/com/mycompany/EntityBeans/Dish.java
@@ -180,7 +180,7 @@ public class Dish implements Serializable {
         return (int)cost;
     }
 
-    public void setCost(double cost) {
+    public void setCost(int cost) {
         this.cost = cost;
     }
